lottowinsv3-backend/requirements.txt
django>=5.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
pymysql>=1.1.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
Pillow>=10.0.0
numpy>=2.2.0
pandas>=2.0.0
scipy>=1.10.0 